Software/system/shuffleboard_1D/script_based_system/digital_or_analogue_system/requirements.txt
PyQt5
brainflow
mne
numpy
paho-mqtt
pyaudio
pyqtgraph
scipy
serial
soundfile
zmq
